#63 [feat] : 냉장고 커스텀 다이얼로그 클래스 생성
DialogFragment 사용
